0.1.44

+++ Utils/Ingredient.cs
@@ -1,7 +1,7 @@
 ﻿namespace Structify;
 
-public class Ingredient(short itemId, int amount)
+public class Ingredient(int itemId, int amount)
 {
-    public short ItemId { get; set; } = itemId;
+    public int ItemId { get; set; } = itemId;
     public int Amount { get; set; } = amount;
 }
